requirements.txt
-i https://pypi.org/simple
absl-py==1.4.0 ; python_version >= '3.6'
anyio==3.6.2 ; python_full_version >= '3.6.2'
astunparse==1.6.3
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
fastapi==0.95.1
flatbuffers==23.5.9
gast==0.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==2.18.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==1.0.0 ; python_version >= '3.6'
google-pasta==0.2.0
grpcio==1.54.2 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
h5py==3.8.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
jax==0.4.10 ; python_version >= '3.8'
keras==2.12.0 ; python_version >= '3.8'
libclang==16.0.0
markdown==3.4.3 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
ml-dtypes==0.1.0 ; python_version >= '3.7'
numpy==1.23.5
oauthlib==3.2.2 ; python_version >= '3.6'
opt-einsum==3.3.0 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
pillow==9.5.0
protobuf==4.23.0 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==1.10.7 ; python_version >= '3.7'
python-multipart==0.0.6
requests==2.30.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
scipy==1.10.1
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.26.1 ; python_version >= '3.7'
tensorboard==2.12.3 ; python_version >= '3.8'
tensorboard-data-server==0.7.0 ; python_version >= '3.7'
tensorflow==2.12.0
tensorflow-estimator==2.12.0 ; python_version >= '3.7'
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0 ; platform_system == 'Windows'
tensorflow-io-gcs-filesystem==0.31.0 ; platform_machine != 'arm64' or platform_system != 'Darwin'
termcolor==2.3.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.22.0
werkzeug==2.3.4 ; python_version >= '3.8'
wheel==0.40.0 ; python_version >= '3.7'
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
